+++ NEWS.md
@@ -4,6 +4,10 @@
 
 - For an input file that contains no code chunks or inline R code expressions, `purl()` should return an empty string instead of the original text from the input (thanks, @jrnold, #1660).
 
+## MAJOR CHANGES
+
+- `knitr::knit()` will try to read the input file with the UTF-8 encoding first. If UTF-8 doesn't work, it will try to read with the system native encoding instead (with a warning). The `encoding` argument of `knitr::knit()` is completely ignored. In the future, only UTF-8 will be supported, and we will stop retrying with the system native encoding. The output file from `knitr::knit()` is always encoded in UTF-8.
+
 # CHANGES IN knitr VERSION 1.22
 
 ## NEW FEATURES